+++ src/functions/uniquify-array.ts
@@ -1,3 +1,3 @@
 export function uniquifyArray<T>(array: T[]): T[] {
-    return array.filter((element, index, self) => index === self.indexOf(element));
+    return [...new Set(array)];
 }