+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Changed
+- Use Set to make an array unique.
+
 ## [1.2.0] - 2019-08-19
 ### Added
 - Ability to give starting number and step amount to generateNumbersArray.