+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!--## [Unreleased]-->
+## [2.0.0] - 2019-11-14
 ### BREAKING CHANGES
 - Dropped support for Node.js 11.
 
@@ -29,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial code.
 - Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/array-utils/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/Ionaru/array-utils/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/Ionaru/array-utils/compare/1.2.0...2.0.0
 [1.2.0]: https://github.com/Ionaru/array-utils/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/Ionaru/array-utils/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/Ionaru/array-utils/compare/971ba24...1.0.0